mlserver/requirements.txt
# General dependencies
numpy==1.24.3
pandas==2.2.3
scikit-learn==1.6.0
matplotlib==3.7.2
seaborn==0.13.2
joblib==1.3.2

# Flask and related dependencies
flask==3.1.0
flask-cors==4.0.1

# Requests and related dependencies
requests==2.32.3
requests-cache==1.2.0
retry-requests==2.0.0

# TensorFlow and related dependencies
tensorflow==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
h5py==3.9.0

# OpenMeteo and related dependencies
openmeteo_requests==1.2.0
openmeteo_sdk<1.11.7

# Other dependencies
ollama==0.4.7
absl-py==1.4.0
APScheduler==3.10.4
astunparse==1.6.3
attrs==25.1.0
blinker==1.9.0
cattrs==23.2.3
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.2.1
flatbuffers>=24.0.0
gast==0.4.0
google-pasta==0.2.0
grpcio==1.69.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.5.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.2
pillow==10.4.0
platformdirs==3.10.0
protobuf>=3.20.3,<5.0.0dev
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
rich==13.7.1
scipy==1.11.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==5.2
url-normalize==1.4.3
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
torch
torchvision
gunicorn
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.15.0

# Ultralytics and related dependencies
ultralytics==8.0.43
ultralyticsplus==0.0.28
scrapy
pathlib
albumentations
